requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
Cython==3.0.11
datasets==2.20.0
dill==0.3.8
docker-pycreds==0.4.0
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
glfw==2.7.0
gym-notices==0.0.8
gymnasium==0.29.1
huggingface-hub==0.24.5
idna==3.7
imageio==2.34.2
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
mujoco-py==2.1.2.14
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
sentry-sdk==2.12.0
setproctitle==1.3.3
setuptools==72.1.0
six==1.16.0
smmap==5.0.1
sympy==1.13.1
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.43.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wandb==0.17.5
xxhash==3.4.1
yarl==1.9.4
gym
scikit-learn
debugpy